gateway/pyproject.toml
# loja/gateway/pyproject.toml

[tool.poetry]
name = "gateway-service"
version = "1.0.0"
description = "API Gateway para a Loja de Microsserviços"
authors = ["Marcos Eduardo"] # Altere se desejar
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
httpx = "^0.27.0"
# pyjwt para decodificar o token
pyjwt = "^2.8.0"
# Pydantic para validação de schemas
pydantic = "^2.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"